+++ Utilities/__init__.py
@@ -1,0 +1,2 @@
+from Fetcher import Fetcher
+from Logger import Logger
