+++ WUBRG/__init__.py
@@ -1,2 +1,2 @@
-from consts import *
-from funcs import *
+from consts import COLORS, COLOR_ALIAS_GROUPS, COLOR_ALIASES, COLOUR_GROUPINGS, COLOR_COMBINATIONS
+from funcs import get_color_string, get_color_identity,get_color_supersets, get_color_subsets
