+++ data_fetching/JSONHandler.py
@@ -4,9 +4,10 @@ from datetime import date, datetime
 import json
 
 from WUBRG import COLOR_COMBINATIONS
+from Utilities import Logger
+from Utilities import Fetcher
+
 from utils.settings import DATA_DIR_LOC, DATA_DIR_NAME
-from utils.Logger import Logger
-from utils.Fetcher import Fetcher
 
 
 class JSONHandler:
