+++ game_metadata/FormatMetadata.py
@@ -2,12 +2,12 @@ from typing import Union, Optional
 from functools import cmp_to_key
 from datetime import date, timedelta
 
-from WUBRG import COLOR_INDEXES
-from utils.settings import SETS, FORMATS, SET_CONFIG
-from game_metadata.CallScryfall import CallScryfall
+from WUBRG.consts import COLOR_INDEXES
 
-from game_metadata.CardManager import CardManager
-from game_metadata.Card import Card
+from utils.settings import SETS, FORMATS, SET_CONFIG
+from CallScryfall import CallScryfall
+from CardManager import CardManager
+from Card import Card
 
 
 class SetMetadata:
